requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.3
django-colorfield
django-cors-headers
django-countries
django-debug-toolbar
django-filter
django-jazzmin
django-rest-swagger
django-templated-mail
djangorestframework==3.13.1
djangorestframework-simplejwt
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
generics==2.0.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.0
MarkupSafe==2.1.1
mysqlclient==2.1.0
oauthlib==3.2.0
openapi-codec==1.3.2
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
#psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyotp==2.6.0
pyparsing==3.0.7
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==2.70.0
termcolor==1.1.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
twilio==7.8.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
validators==0.18.2
whitenoise==6.0.0
wsproto==1.1.0
